+++ json_annotation/lib/src/json_key.dart
@@ -47,12 +47,25 @@ class JsonKey {
   ///
   /// If `null` (the default) or `false`, the field will be considered for
   /// serialization.
+  ///
+  /// This field is DEPRECATED use [includeFromJson] and [includeToJson]
+  /// instead.
   @Deprecated(
     'Use `includeFromJson` and `includeToJson` with a value of `false` '
     'instead.',
   )
   final bool? ignore;
 
+  /// Used to force a field to be included (or excluded) when decoding a object
+  /// from JSON.
+  ///
+  /// `null` (the default) means the field will be handled with the default
+  /// semantics that take into account if it's private or if it can be cleanly
+  /// round-tripped to-from JSON.
+  ///
+  /// `true` means the field should always be decoded, even if it's private.
+  ///
+  /// `false` means the field should never be decoded.
   final bool? includeFromJson;
 
   /// Whether the generator should include fields with `null` values in the
@@ -72,6 +85,16 @@ class JsonKey {
   /// same field, an exception will be thrown during code generation.
   final bool? includeIfNull;
 
+  /// Used to force a field to be included (or excluded) when encoding a object
+  /// to JSON.
+  ///
+  /// `null` (the default) means the field will be handled with the default
+  /// semantics that take into account if it's private or if it can be cleanly
+  /// round-tripped to-from JSON.
+  ///
+  /// `true` means the field should always be encoded, even if it's private.
+  ///
+  /// `false` means the field should never be encoded.
   final bool? includeToJson;
 
   /// The key in a JSON map to use when reading and writing values corresponding