+++ json_serializable/CHANGELOG.md
@@ -1,7 +1,8 @@
-## 6.6.0-dev
+## 6.6.0
 
 - Support for `JsonKey.includeFromJson` and `JsonKey.includeToJson`.
 - Support `JsonEnum.valueField` being set with `'index'`.
+- Support `JsonSerializable.createPerFieldToJson`.
 - Require Dart SDK `>=2.18.0`.
 - Require `analyzer: ^5.2.0`
 - Require `json_annotation: '>=4.8.0 <4.9.0'`